refactor

+++ csrc/device_lower/pass/double_buffer.h
@@ -217,7 +217,7 @@ class DoubleBufferInfo {
 
   //! Get the number of circular buffer stages for the given axis,
   //!  the number of stages will be 2 in the case of double buffer loop.
-  unsigned int getStageDepthFor(IterDomain* circular_buffered_id);
+  int64_t getStageDepthFor(IterDomain* circular_buffered_id);
 
  private:
   TvInfo& getTvInfo(const TensorView* tv);
